feat: 내비게이션 Scaffold로 HomePage 구조 변경
